% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antiRetroviralIntervals.R
\name{antiRetroviralIntervals}
\alias{antiRetroviralIntervals}
\title{Intervals from anti-retroviral access data}
\usage{
antiRetroviralIntervals()
}
\value{
Data frame
}
\description{
Generates a data frame with the interval estimates from
\link{antiRetroviral} in the format described in
\link{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
}
