% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottomSet.R
\name{bottomSet}
\alias{bottomSet}
\title{Compute the k-bottom set for a set of intervals}
\usage{
bottomSet(intervals, k, rankIntervals = FALSE)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}

\item{k}{cutoff for inclusion}

\item{rankIntervals}{true if the intervals are rank intervals and false
otherwise}
}
\value{
Indices of intervals in the k-bottom set.
}
\description{
Suppose that we select one point from each of a set of n intervals and rank
them.  The k-bottom set is the set of intervals whose points can have a rank
of n + 1 - k or higher.
}
\details{
See section 7 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
bottomSet(intervals, 2)
}
\references{
\insertAllCited{}
}
