% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{dual}
\alias{dual}
\title{Generate the dual of an interval order}
\usage{
dual(intervals)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}
}
\value{
data frame
}
\description{
Given a representation of an interval order, this function generates a
representation of the dual order.
}
\details{
See section 2 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
dual(intervals)
}
\references{
\insertAllCited{}
}
