% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsToMatrix.R
\name{intervalsToMatrix}
\alias{intervalsToMatrix}
\title{Matrix representation of the order generated by a set of intervals}
\usage{
intervalsToMatrix(intervals, strict = FALSE, binary = FALSE)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}

\item{strict}{is this <= or <?}

\item{binary}{output is coded as 0/1 if TRUE and FALSE/TRUE otherwise}
}
\value{
A boolean matrix.  If strict is set to TRUE, the (i, j)th entry is
intervals[i, 'right'] < intervals[j, 'left'].  If strict is set to false, <=
is used in place of <.
}
\description{
Represent the order generated by a set of intervals as a Boolean matrix.
This is a common input format for programs that operate on partial orders and
can be used to compute the distance between those orders.
}
\details{
See sections 3 and 6 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()[2:8, ]
intervalsToMatrix(intervals)
intervalsToMatrix(intervals, strict = TRUE)
intervalsToMatrix(intervals, binary = TRUE)
}
\references{
\insertAllCited{}
}
