% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCompatible.R
\name{isCompatible}
\alias{isCompatible}
\title{Test whether a ranking is compatible with a set of intervals}
\usage{
isCompatible(ranking, intervals)
}
\arguments{
\item{ranking}{ranking vector (see \link{orderRanks} for format)}

\item{intervals}{data frame (see \link{orderRanks} for
the required format)}
}
\value{
boolean
}
\description{
This method determines whether there is some set of points contained in a set
of intervals whose ranking is equal to a given ranking.
}
\details{
See section 8 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()[2:8, ]
isCompatible(1:7, intervals)
isCompatible(7:1, intervals)
}
\references{
\insertAllCited{}
}
