% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessThan.R
\name{lessThan}
\alias{lessThan}
\title{Compare intervals}
\usage{
lessThan(intervals, i, j)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}

\item{i}{row index of left-hand side of inequality}

\item{j}{row index of right-hand side of inequaltiy}
}
\value{
Boolean value
}
\description{
Given a data frame representing a set of intervals, return true if row i is
less than row j under the order generated by the intervals and false
otherwise.
}
\details{
See section 3 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
lessThan(intervals, 2, 3)
lessThan(intervals, 1, 10)
}
\references{
\insertAllCited{}
}
