% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingDistance.R
\name{rankingDistance}
\alias{rankingDistance}
\title{Calculate the distance between the orders generated by two rankings}
\usage{
rankingDistance(r1, r2, normalize = FALSE)
}
\arguments{
\item{r1}{ranking vector (see \link{orderRanks} for format)}

\item{r2}{ranking vector}

\item{normalize}{boolean if TRUE, normalize the distance to [0, 1]}
}
\value{
numeric measure of distance
}
\description{
Given two rankings, this method determines the distance between the orders
they represent.
}
\details{
See section 6 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
rankingDistance(1:7, 1:7)
rankingDistance(1:7, 7:1, TRUE)
rankingDistance(1:7, 7:1)
rankingDistance(1:7, c(7, 1:6))
rankingDistance(1:7, c(2, 1, 3:7))
}
\references{
\insertAllCited{}
}
