% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingToMatrix.R
\name{rankingToMatrix}
\alias{rankingToMatrix}
\title{Matrix representation of the order generated by a ranking}
\usage{
rankingToMatrix(ranking, strict = FALSE, binary = FALSE)
}
\arguments{
\item{ranking}{rank vector (see \link{orderRanks} for format)}

\item{strict}{is this <= or <?}

\item{binary}{output is coded as 0/1 if TRUE and FALSE/TRUE otherwise}
}
\value{
A boolean matrix.  If strict is set to TRUE, the (i, j)th entry is
ranking[i] < ranking[j].  If strict is set to false, <= is used in place
of <.
}
\description{
Represent the order generated by a ranking as a Boolean matrix.  This is a
common input format for programs that operate on partial orders and can be
used to compute the distance between those orders.
}
\details{
See sections 3 and 6 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
r1 <- rank(antiRetroviralIntervals()[2:8, 'left'], ties.method = 'min')
rankingToMatrix(r1)
rankingToMatrix(r1, strict = TRUE)
rankingToMatrix(r1, binary = TRUE)

r2 <- rank(antiRetroviralIntervals()[2:8, 'right'], ties.method = 'min')
rankingToMatrix(r2)
rankingToMatrix(r2, strict = TRUE)
rankingToMatrix(r2, binary = TRUE)
}
\references{
\insertAllCited{}
}
