% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIntervals.R
\name{rankIntervals}
\alias{rankIntervals}
\title{Generate rank intervals for a set of intervals}
\usage{
rankIntervals(intervals)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}
}
\value{
data frame
}
\description{
Given a set of interval estimates, this method generates a set of intervals
that describe which ranks the estimated parameters can have if they are
contained in the interval estimates.
}
\details{
See section 4 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
rankIntervals(intervals)
}
\references{
\insertAllCited{}
}
