% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankUncertainty.R
\name{rankUncertainty}
\alias{rankUncertainty}
\title{Compute the conditional uncertainty in a set of rankings}
\usage{
rankUncertainty(
  intervals,
  symmetrize = TRUE,
  normalize = TRUE,
  rankIntervals = FALSE
)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}

\item{symmetrize}{if true, compute the symmetrized uncertainty}

\item{normalize}{if true, compute the normalized uncertainty}

\item{rankIntervals}{true if the intervals are rank intervals and false
otherwise}
}
\value{
Numeric value measuring the conditional uncertainty in the ranks
}
\description{
Computes a measure of the uncertainty in the estimated ranks of a set of
parameters based on the uncertainty in the parameter estimates as represented
by a set of interval estimates.
}
\details{
See section 7 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
rankUncertainty(intervals)
}
\references{
\insertAllCited{}
}
