% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleIntervalOrderRanks.R
\name{sampleIntervalOrderRanks}
\alias{sampleIntervalOrderRanks}
\title{Compute the sample interval order ranks}
\usage{
sampleIntervalOrderRanks(intervals, rankIntervals = FALSE)
}
\arguments{
\item{intervals}{data frame (see \link{orderRanks} for
the required format)}

\item{rankIntervals}{true if the intervals are rank intervals and false
otherwise}
}
\value{
ranking vector (see \link{orderRanks} for format)
}
\description{
Computes the sample interval order ranks for a set of unknown parameters
from a collection of interval estimates for those parameters.
}
\details{
See section 6 of \insertCite{rising2023orderRanks;textual}{orderRanks}.
}
\examples{
intervals <- antiRetroviralIntervals()
sampleIntervalOrderRanks(intervals)
}
\references{
\insertAllCited{}
}
